backend/pyproject.toml
[project]
name = "ttg-backend"
version = "0.1.0"
description = "TTG Backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.16",
    "fastapi[standard]>=0.115.12",
    "gitingest>=0.1.4",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-pinecone>=0.2.0",
    "langchain-text-splitters>=0.3.0",
    "pinecone-client>=5.0.0",
    "flashrank>=0.2.0",
    "rank-bm25>=0.2.2",
    "websockets>=15.0.1",
    "aiosqlite>=0.20.0",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "nanoid>=2.0.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.2",
    "types-nanoid>=2.0.0.20240601",
]


[tool.pytest.ini_options]
pythonpath = [
    ".",
]
asyncio_mode = "strict"
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"
